+++ src/components/common/IdolImage.jsx
