+++ src/components/common/IdolImage.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/color.scss";
+@import "../../styles/color.scss";
 
 .imgBorder {
   width: var(--size);
